requirement.txt
streamlit==1.32.0
google-cloud-aiplatform==1.45.0
python-dotenv==1.0.1